added panic catcher in viewcontext::CallView
